Add change function: avoids repeated change
The change(...) function takes a tk element (meant to be a label for now), a parameter (e.g., 'text'), and the new value to assign. If the value is the same as the present value, it doesn't change the parameter. This way, the label display is refreshed less often, and rendering becomes faster.